+++ packages/page-navigation/src/types/StoreProps.ts
@@ -6,8 +6,11 @@
  * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
  */
 
+import type { PdfJs } from '@react-pdf-viewer/core';
+
 export interface StoreProps {
     currentPage?: number;
+    doc?: PdfJs.PdfDocument;
     numberOfPages?: number;
     jumpToPage?(pageIndex: number): void;
 }